+++ lib/bones_api.dart
@@ -243,6 +243,7 @@ export 'src/bones_api_security.dart';
 export 'src/bones_api_session.dart';
 export 'src/bones_api_sql_builder.dart';
 export 'src/bones_api_types.dart';
+export 'src/bones_api_utils.dart';
 export 'src/bones_api_utils_arguments.dart';
 export 'src/bones_api_utils_collections.dart';
 export 'src/bones_api_utils_httpclient.dart';