+++ lib/src/bones_api_entity_annotation.dart
@@ -1,4 +1,3 @@
-import 'package:bones_api/bones_api.dart';
 import 'package:collection/collection.dart';
 import 'package:meta/meta_meta.dart';
 import 'package:statistics/statistics.dart';